Create LinkedList
A linked list is a type of linear data structure similar to arrays. It is a collection of nodes that are linked with each other. A node contains two things first is data and second is a link that connects it with another node. Our first node is where head points and we can access all the elements of the linked list using the head.